Added database component to invalidate password reset token after it has been used

+++ service/token-verification-service.js
@@ -1,6 +1,10 @@
-
 const VerificationToken = require('../models/VerificationToken');
-const { createJWTEmail, verifyEmailToken } = require('../utils/tokenUtils');
+const {
+  createJWTEmail,
+  verifyEmailToken,
+  verifyResetToken,
+  createJWTPasswordReset,
+} = require('../utils/tokenUtils');
 
 const createToken = async (userId, email) => {
   const token = createJWTEmail({ email });
@@ -8,19 +12,39 @@ const createToken = async (userId, email) => {
   return token;
 };
 
-const verifyToken = async (userId, token) => {
+const verifyToken = async (userId, token, tokenType) => {
   const storedToken = await VerificationToken.findOne({ userId, token });
   if (!storedToken) {
     return { valid: false, reason: 'expired' };
   }
 
-  try {
-    verifyEmailToken(token);
-    return { valid: true };
-  } catch (err) {
-    await VerificationToken.deleteOne({ userId, token });
-    return { valid: false, reason: 'invalid' };
+  if (tokenType === 'email') {
+    try {
+      verifyEmailToken(token);
+      return { valid: true };
+    } catch (err) {
+      await VerificationToken.deleteOne({ userId, token });
+      return { valid: false, reason: 'invalid' };
+    }
+  } else if (tokenType === 'password') {
+    try {
+      const tokenStatus = verifyResetToken(token);
+      if (tokenStatus.exp < Math.round(Date.now() / 1000)) {
+        await VerificationToken.deleteOne({ userId, token });
+        return { valid: false, reason: 'expired' };
+      }
+      return { valid: true };
+    } catch (err) {
+      await VerificationToken.deleteOne({ userId, token });
+      return { valid: false, reason: 'invalid' };
+    }
   }
+};
+
+const createPasswordToken = async (userId, email) => {
+  const token = createJWTPasswordReset({ email });
+  await VerificationToken.create({ userId, token });
+  return token;
 };
 
 const invalidateTokens = async (userId) => {
@@ -31,4 +55,5 @@ module.exports = {
   createToken,
   verifyToken,
   invalidateTokens,
+  createPasswordToken,
 };
